Check inventory port before default port 22
The rational being that it should only need to fallback on the
first run. From then on it should always be reachable on the
inventory port. So now we will only check port 22 if the inventory
port is unreachable. This will avoid wasting time waiting for
port 22 after the first run.